mlcore/pyproject.toml
[project]
name = "mlcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.2",
    "expects>=0.9.0",
    "notebook>=7.5.0",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.3.1",
    "pyarrow>=22.0.0",
    "pymc>=5.26.1",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
